基本布局和简单数据加载

+++ app/src/main/java/com/yyl/obstetriciansassistant/view/fragments/CaseFragment.kt
@@ -1,20 +1,21 @@
 package com.yyl.obstetriciansassistant.view.fragments
 
+import android.app.AlertDialog
 import android.os.Bundle
 import android.support.v4.app.Fragment
 import android.support.v7.widget.LinearLayoutManager
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.yyl.obstetriciansassistant.R
+import com.yyl.obstetriciansassistant.*
 import com.yyl.obstetriciansassistant.beans.Case
 import com.yyl.obstetriciansassistant.model.CaseModelImpl
-import com.yyl.obstetriciansassistant.toast
+import com.yyl.obstetriciansassistant.view.activities.DetailActivity
 import com.yyl.obstetriciansassistant.view.adapter.CaseAdapter
 import kotlinx.android.synthetic.main.fragment_case.*
 
 class CaseFragment : Fragment() {
-    private var list: MutableList<Case>? = arrayListOf()
+    private var list = arrayListOf<Case>()
     private val caseModel = CaseModelImpl()
     private lateinit var adapter: CaseAdapter
 
@@ -29,29 +30,37 @@ class CaseFragment : Fragment() {
     }
 
     private fun initCase() {
-        list=caseModel.getCases()
-        adapter = CaseAdapter(list!!)
+        list = caseModel.getCases()
+        adapter = CaseAdapter(list)
         adapter.onItemClickListener = object : CaseAdapter.OnItemClickListener {
             override fun onItemClick(v: View, position: Int) {
-                toast("item click")
+                activity!!.jump2Activity(activity!!, DetailActivity::class.java, CASE, list[position])
             }
 
             override fun onItemEditClick(v: View, position: Int) {
                 toast("edit click $position")
+                activity!!.jump2Activity(activity!!, DetailActivity::class.java, CASE, CHANGE,list[position])
             }
 
             override fun onItemDeleteClick(v: View, position: Int) {
-                if (caseModel.deleteCase(position)){
-                    list!!.removeAt(position)
-                    adapter.notifyDataSetChanged()
-                }else{
-                    toast("delete failed $position")
-                }
-            }
+                AlertDialog.Builder(activity!!)
+                    .setTitle("提示")
+                    .setMessage("您正在删除${list[position].name}的病例，请谨慎操作")
+                    .setPositiveButton("删除") { _, _ ->
+                        if (caseModel.deleteCase(position)) {
+                            list.removeAt(position)
+                            adapter.notifyDataSetChanged()
+                        } else {
+                            toast("delete failed $position")
+                        }
+                    }
+                    .setNegativeButton("取消", null)
+                    .create()
+                    .show()
 
+            }
         }
         case_list_view.layoutManager = LinearLayoutManager(activity!!, LinearLayoutManager.VERTICAL, false)
-        //case_list_view.addItemDecoration(DividerItemDecoration(activity,DividerItemDecoration.VERTICAL))
-        case_list_view.adapter=adapter
+        case_list_view.adapter = adapter
     }
 }
